requirements.txt
###
# KMS Service is an AWS Lambda interface to interact with AWS KMS Key Encryption
# service.
# 
# Copyright (C) 2018-2019  Asurion, LLC
#
# KMS Service is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# KMS Service is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with KMS Service.  If not, see <https://www.gnu.org/licenses/>.
###

pycrypto==2.6.1